Handle notes commands in daemon
