+++ crates/roux-runtime/Cargo.toml
@@ -14,6 +14,7 @@ portable-pty = "0.9"
 roux-core = { path = "../roux-core" }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+serde_yaml = "0.9"
 sha2 = "0.10"
 specta = { version = "=2.0.0-rc.24", features = ["derive", "serde_json"] }
 thiserror = "2"
